requirements.txt
# 大规模成像星座任务规划框架 - 依赖

# 核心科学计算
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# 可视化
matplotlib>=3.4.0

# 轨道传播 (Mock STK 用)
sgp4>=2.20

# 地理计算
shapely>=1.8.0
pyproj>=3.2.0

# 配置和序列化
pyyaml>=6.0
dataclasses-json>=0.5.0

# 进度条
tqdm>=4.62.0

# 测试框架
pytest>=7.0.0
pytest-cov>=3.0.0

# ======= Windows 额外依赖 =======
# pywin32>=300  # 仅 Windows 需要，用于 STK COM 接口
